+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "hotel-reservation"
-version = "0.0.8"
+version = "0.2.6"
 description = "Hotel reservations project"
 requires-python = ">=3.11, <3.12"
 dependencies = ["mlflow==2.17.0",
@@ -16,8 +16,7 @@ dependencies = ["mlflow==2.17.0",
                 "scikit-learn==1.5.2",
                 "lightgbm==4.5.0",
                 "scipy==1.14.1",
-                "databricks-feature-engineering==0.6",
-                "databricks-feature-lookup==1.2.0",
+                "toml==0.10.2",
                 "databricks-sdk==0.32.0",
                 "pydantic==2.9.2"
                 ]
